Make the change more localised

+++ point.go
@@ -5,17 +5,22 @@ type Point struct {
 }
 
 // Move the point by one step (dir is expected to be 0, 1, 2 or 3) and clamp it
-// to the confines of the world.
-func (p Point) MoveAndClamp(dir uint) Point {
+// to the confines of the world. It is assumed that the initial value of p is
+// within the confines of the world.
+func (p Point) MoveAndClamp(dir int) Point {
 	a := dir & 1
 	b := dir >> 1
-	x := uint(p.X) + a - b
-	y := uint(p.Y) + a + b - 1
-	if x < worldWidth {
-		p.X = int(x)
+	x := p.X + a - b
+	y := p.Y + a + b - 1
+	if uint(x) < worldWidth {
+		// We get to this branch if and only if 0 <= x < worldwidth as if x < 0
+		// then uint(x) wraps around to a very big number (bigger than
+		// worldWidth)
+		p.X = x
 	}
-	if y < worldHeight {
-		p.Y = int(y)
+	if uint(y) < worldHeight {
+		// As above we get to this branch if and only if 0 <= y < worldHeight
+		p.Y = y
 	}
 	return p
 }
